live update rows
instead of updating after end of drag.
This might not be the most performant solution, since it triggers a
re-render of the list each time a element changes position. But since
this kind of app mostly won't have more then 30 players within a
registriation list this will be ok.

also i introduced a drag handler, to increase visibility of the dragging
feature